# generated/api/notePurchaseAgreement.pb.gw.go
// Code generated by protoc-gen-grpc-gateway. DO NOT EDIT.
// source: proto/api/notePurchaseAgreement.proto

/*
Package notepurchaseagreement is a reverse proxy.

It translates gRPC into RESTful JSON APIs.
*/
package notepurchaseagreement

import (
	"context"
	"io"
	"net/http"

	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
	"github.com/grpc-ecosystem/grpc-gateway/v2/utilities"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/grpclog"
	"google.golang.org/grpc/metadata"
	"google.golang.org/grpc/status"
	"google.golang.org/protobuf/proto"
)

// Suppress "imported and not used" errors
var _ codes.Code
var _ io.Reader
var _ status.Status
var _ = runtime.String
var _ = utilities.NewDoubleArray
var _ = metadata.Join

func request_NotePurchaseAgreementService_GetNotePurchaseAgreements_0(ctx context.Context, marshaler runtime.Marshaler, client NotePurchaseAgreementServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq GetNotePurchaseAgreementsRequest
	var metadata runtime.ServerMetadata

	msg, err := client.GetNotePurchaseAgreements(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_NotePurchaseAgreementService_GetNotePurchaseAgreements_0(ctx context.Context, marshaler runtime.Marshaler, server NotePurchaseAgreementServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq GetNotePurchaseAgreementsRequest
	var metadata runtime.ServerMetadata

	msg, err := server.GetNotePurchaseAgreements(ctx, &protoReq)
	return msg, metadata, err

}

var (
	filter_NotePurchaseAgreementService_GetNotePurchaseAgreement_0 = &utilities.DoubleArray{Encoding: map[string]int{}, Base: []int(nil), Check: []int(nil)}
)

func request_NotePurchaseAgreementService_GetNotePurchaseAgreement_0(ctx context.Context, marshaler runtime.Marshaler, client NotePurchaseAgreementServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq GetNotePurchaseAgreementRequest
	var metadata runtime.ServerMetadata

	if err := req.ParseForm(); err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}
	if err := runtime.PopulateQueryParameters(&protoReq, req.Form, filter_NotePurchaseAgreementService_GetNotePurchaseAgreement_0); err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := client.GetNotePurchaseAgreement(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_NotePurchaseAgreementService_GetNotePurchaseAgreement_0(ctx context.Context, marshaler runtime.Marshaler, server NotePurchaseAgreementServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq GetNotePurchaseAgreementRequest
	var metadata runtime.ServerMetadata

	if err := req.ParseForm(); err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}
	if err := runtime.PopulateQueryParameters(&protoReq, req.Form, filter_NotePurchaseAgreementService_GetNotePurchaseAgreement_0); err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := server.GetNotePurchaseAgreement(ctx, &protoReq)
	return msg, metadata, err

}

var (
	filter_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0 = &utilities.DoubleArray{Encoding: map[string]int{}, Base: []int(nil), Check: []int(nil)}
)

func request_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0(ctx context.Context, marshaler runtime.Marshaler, client NotePurchaseAgreementServiceClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq SaveNotePurchaseAgreementRequest
	var metadata runtime.ServerMetadata

	if err := req.ParseForm(); err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}
	if err := runtime.PopulateQueryParameters(&protoReq, req.Form, filter_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0); err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := client.SaveNotePurchaseAgreement(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0(ctx context.Context, marshaler runtime.Marshaler, server NotePurchaseAgreementServiceServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq SaveNotePurchaseAgreementRequest
	var metadata runtime.ServerMetadata

	if err := req.ParseForm(); err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}
	if err := runtime.PopulateQueryParameters(&protoReq, req.Form, filter_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0); err != nil {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := server.SaveNotePurchaseAgreement(ctx, &protoReq)
	return msg, metadata, err

}

// RegisterNotePurchaseAgreementServiceHandlerServer registers the http handlers for service NotePurchaseAgreementService to "mux".
// UnaryRPC     :call NotePurchaseAgreementServiceServer directly.
// StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
// Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterNotePurchaseAgreementServiceHandlerFromEndpoint instead.
func RegisterNotePurchaseAgreementServiceHandlerServer(ctx context.Context, mux *runtime.ServeMux, server NotePurchaseAgreementServiceServer) error {

	mux.Handle("GET", pattern_NotePurchaseAgreementService_GetNotePurchaseAgreements_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/notepurchaseagreement.NotePurchaseAgreementService/GetNotePurchaseAgreements", runtime.WithHTTPPathPattern("/v1/notepurchaseagreements"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_NotePurchaseAgreementService_GetNotePurchaseAgreements_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_NotePurchaseAgreementService_GetNotePurchaseAgreements_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("GET", pattern_NotePurchaseAgreementService_GetNotePurchaseAgreement_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/notepurchaseagreement.NotePurchaseAgreementService/GetNotePurchaseAgreement", runtime.WithHTTPPathPattern("/v1/notepurchaseagreement"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_NotePurchaseAgreementService_GetNotePurchaseAgreement_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_NotePurchaseAgreementService_GetNotePurchaseAgreement_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("POST", pattern_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/notepurchaseagreement.NotePurchaseAgreementService/SaveNotePurchaseAgreement", runtime.WithHTTPPathPattern("/v1/notepurchaseagreement"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

// RegisterNotePurchaseAgreementServiceHandlerFromEndpoint is same as RegisterNotePurchaseAgreementServiceHandler but
// automatically dials to "endpoint" and closes the connection when "ctx" gets done.
func RegisterNotePurchaseAgreementServiceHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
	conn, err := grpc.Dial(endpoint, opts...)
	if err != nil {
		return err
	}
	defer func() {
		if err != nil {
			if cerr := conn.Close(); cerr != nil {
				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
			}
			return
		}
		go func() {
			<-ctx.Done()
			if cerr := conn.Close(); cerr != nil {
				grpclog.Infof("Failed to close conn to %s: %v", endpoint, cerr)
			}
		}()
	}()

	return RegisterNotePurchaseAgreementServiceHandler(ctx, mux, conn)
}

// RegisterNotePurchaseAgreementServiceHandler registers the http handlers for service NotePurchaseAgreementService to "mux".
// The handlers forward requests to the grpc endpoint over "conn".
func RegisterNotePurchaseAgreementServiceHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
	return RegisterNotePurchaseAgreementServiceHandlerClient(ctx, mux, NewNotePurchaseAgreementServiceClient(conn))
}

// RegisterNotePurchaseAgreementServiceHandlerClient registers the http handlers for service NotePurchaseAgreementService
// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "NotePurchaseAgreementServiceClient".
// Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "NotePurchaseAgreementServiceClient"
// doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
// "NotePurchaseAgreementServiceClient" to call the correct interceptors.
func RegisterNotePurchaseAgreementServiceHandlerClient(ctx context.Context, mux *runtime.ServeMux, client NotePurchaseAgreementServiceClient) error {

	mux.Handle("GET", pattern_NotePurchaseAgreementService_GetNotePurchaseAgreements_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/notepurchaseagreement.NotePurchaseAgreementService/GetNotePurchaseAgreements", runtime.WithHTTPPathPattern("/v1/notepurchaseagreements"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_NotePurchaseAgreementService_GetNotePurchaseAgreements_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_NotePurchaseAgreementService_GetNotePurchaseAgreements_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("GET", pattern_NotePurchaseAgreementService_GetNotePurchaseAgreement_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/notepurchaseagreement.NotePurchaseAgreementService/GetNotePurchaseAgreement", runtime.WithHTTPPathPattern("/v1/notepurchaseagreement"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_NotePurchaseAgreementService_GetNotePurchaseAgreement_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_NotePurchaseAgreementService_GetNotePurchaseAgreement_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("POST", pattern_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/notepurchaseagreement.NotePurchaseAgreementService/SaveNotePurchaseAgreement", runtime.WithHTTPPathPattern("/v1/notepurchaseagreement"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	return nil
}

var (
	pattern_NotePurchaseAgreementService_GetNotePurchaseAgreements_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1", "notepurchaseagreements"}, ""))

	pattern_NotePurchaseAgreementService_GetNotePurchaseAgreement_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1", "notepurchaseagreement"}, ""))

	pattern_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1", "notepurchaseagreement"}, ""))
)

var (
	forward_NotePurchaseAgreementService_GetNotePurchaseAgreements_0 = runtime.ForwardResponseMessage

	forward_NotePurchaseAgreementService_GetNotePurchaseAgreement_0 = runtime.ForwardResponseMessage

	forward_NotePurchaseAgreementService_SaveNotePurchaseAgreement_0 = runtime.ForwardResponseMessage
)