# generated/openapi.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/api/auth.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "AuthenticationService"
    },
    {
      "name": "NotePurchaseAgreementService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/authenticate-user": {
      "post": {
        "summary": "Authenticates a user by username and password",
        "description": "Authenticates a user by username and password",
        "operationId": "AuthenticationService_AuthenticateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authAuthenticateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/v1/notepurchaseagreement": {
      "get": {
        "summary": "Get Note Purchase Agreement Document",
        "description": "Gets a note purchase agreement filled out with customer data.",
        "operationId": "NotePurchaseAgreementService_GetNotePurchaseAgreement",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/notepurchaseagreementGetNotePurchaseAgreementResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "Payload.Id",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "NotePurchaseAgreement"
        ]
      },
      "post": {
        "summary": "Save Note Purchase Agreement Document",
        "description": "Saves a note purchase agreement filled out with customer data.",
        "operationId": "NotePurchaseAgreementService_SaveNotePurchaseAgreement",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/notepurchaseagreementSaveNotePurchaseAgreementResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "NotePurchaseAgreement"
        ]
      }
    },
    "/v1/notepurchaseagreements": {
      "get": {
        "summary": "Get All Note Purchase Agreement Documents",
        "description": "Gets note purchase agreements with limited data.",
        "operationId": "NotePurchaseAgreementService_GetNotePurchaseAgreements",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/notepurchaseagreementGetNotePurchaseAgreementsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "NotePurchaseAgreement"
        ]
      }
    }
  },
  "definitions": {
    "authAuthenticateUserResponse": {
      "type": "object",
      "properties": {
        "authenticated": {
          "type": "boolean"
        },
        "errorMessage": {
          "type": "string"
        }
      }
    },
    "authUser": {
      "type": "object",
      "properties": {
        "Id": {
          "type": "string"
        },
        "EmailAddress": {
          "type": "string"
        },
        "Password": {
          "type": "string"
        }
      }
    },
    "notepurchaseagreementGetNotePurchaseAgreementResponse": {
      "type": "object",
      "properties": {
        "payload": {
          "$ref": "#/definitions/notepurchaseagreementNotePurchaseAgreementRecord"
        }
      }
    },
    "notepurchaseagreementGetNotePurchaseAgreementsResponse": {
      "type": "object",
      "properties": {
        "payload": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/notepurchaseagreementRecordCollection"
          }
        }
      }
    },
    "notepurchaseagreementNotePurchaseAgreementRecord": {
      "type": "object",
      "properties": {
        "Id": {
          "type": "string"
        },
        "FirstName": {
          "type": "string"
        },
        "LastName": {
          "type": "string"
        },
        "DateOfBirth": {
          "type": "string"
        },
        "HomeAddress": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/authUser"
        },
        "PhoneNumber": {
          "type": "string"
        },
        "SocialSecurity": {
          "type": "string"
        },
        "FundsCommitted": {
          "type": "string",
          "format": "uint64"
        },
        "FileContent": {
          "type": "string",
          "format": "byte"
        },
        "CreatedOn": {
          "type": "string"
        }
      }
    },
    "notepurchaseagreementRecordCollection": {
      "type": "object",
      "properties": {
        "Payload": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/notepurchaseagreementRecordRequestPayload"
          }
        }
      }
    },
    "notepurchaseagreementRecordRequestPayload": {
      "type": "object",
      "properties": {
        "Id": {
          "type": "string"
        }
      }
    },
    "notepurchaseagreementRecordResultPayload": {
      "type": "object",
      "properties": {
        "Id": {
          "type": "string"
        },
        "CreatedOn": {
          "type": "string"
        }
      }
    },
    "notepurchaseagreementSaveNotePurchaseAgreementResponse": {
      "type": "object",
      "properties": {
        "payload": {
          "$ref": "#/definitions/notepurchaseagreementRecordResultPayload"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}